refactor(learning): split coaching-prose generation out of reasonTagging (R4)
`learning/reasonTagging.ts` was 1344 lines doing two jobs: classify a move
(features → reason → concept tags → risk flags) and write it up
(`SHORT_EXPLANATION_TEMPLATES` + `generateShortExplanation` +
`generateLongExplanation`, ~475 lines of table + copy-heavy branching).

**Finding correction:** R4 as filed said "move the copy tables" and expected
`reasonExplanationTemplates.ts`. In reality there's one 61-line table and two
copy-heavy *functions* — the honest split is the whole prose concern, into
`reasonExplanations.ts`. `reasonTagging.ts` re-exports the two functions so its
public surface is unchanged; `tagMove` is the only caller and imports them.

Pure move, zero behaviour change — the existing `reasonTagging.test.ts` +
`moveAnalysis` tests cover it.

`reasonTagging.ts` 1344 → 875 lines, which drops it under the `max-lines`
budget — the client lint warning count falls 51 → 50, and `--max-warnings` is
tightened to 50 to hold the ground (per REFACTOR_OPPORTUNITIES §4: the budget is
the ratchet).

Local: `tsc -b`, `lint` 50/50, `lint:hooks`, `check:architecture` 20/20,
`check:deps`, client `test:all` 218/1632, server vitest all shards.
