+++ client/package.json
@@ -27,7 +27,7 @@
     "test:smoke:sockets:stress": "SMOKE_REPEAT=10 node scripts/socketSmoke.mjs",
     "soak:mp-private-authority": "node scripts/mpPrivateAuthoritySoak.mjs",
     "ci:bot": "npm run test:bot && npm run build",
-    "lint": "ESLINT_USE_FLAT_CONFIG=false eslint src --ext .ts,.tsx --max-warnings 51",
+    "lint": "ESLINT_USE_FLAT_CONFIG=false eslint src --ext .ts,.tsx --max-warnings 50",
     "lint:hooks": "ESLINT_USE_FLAT_CONFIG=false eslint src --ext .ts,.tsx --no-eslintrc --config .eslintrc.hooks.json --max-warnings 0",
     "lint:css": "stylelint \"src/**/*.css\"",
     "check:deps": "depcruise src --config .dependency-cruiser.json --output-type text",
